refactor: убрать лишнее
