+++ src/components/Burger/Burger.css
@@ -18,7 +18,7 @@
   bottom: 0;
   max-width: 520px;
   width: 100%;
-  background-color: white;
+  background-color: #FFFFFF;
   list-style: none;
   margin: 0;
   padding: 144px 0 52px;
@@ -75,7 +75,7 @@
   display: inline-block;
   width: 30px;
   height: 30px;
-  background-color: rgba(249, 249, 249, 1);
+  background-color: #FFFFFF;
   background-repeat: no-repeat;
   background-position: center;
   border-radius: 4px;
