+++ src/components/Footer/Footer.css
@@ -7,23 +7,26 @@
     justify-content: flex-end;
     gap: 20px;
     flex: 0 0 auto;
-
 }
+
 .footer__items{
     font-size: 13px;
     display: flex;
     flex-direction: row;
     justify-content: space-between;
 }
+
 .footer__item{
     font-size: 13px;
     margin: 0 0 20px;
 }
+
 .footer__links{
     display: flex;
     flex-direction: row;
     gap: 20px;
 }
+
 .footer__title{
     color: #A0A0A0;
     margin: 0;
@@ -32,14 +35,17 @@
     border-bottom: 1px solid #E8E8E8;
     padding-bottom: 21px;
 }
+
 .footer__copyright{
     color: #000000;
 }
+
 @media screen and (max-width: 768px) {
     .footer{
         background-color: #FFFFFF;
         padding: 0 30px 0;
     }
+
     .footer__title{
         max-width: 708px;
     }
@@ -50,27 +56,32 @@
         gap: 31px;
         min-height: 253px
     }
+
     .footer__title{
         font-size: 12px;
         max-width: 300px;
         align-self: center;
     }
+
     .footer__items{
         flex-direction: column-reverse;
         align-items:center;
         margin: 0;
         margin-bottom: 22px;
     }
+
     .footer__item{
         font-size: 12px;
         margin: 0;
     }
+
     .footer__links{
         flex-direction: column;
         gap: 13px;
         align-items:center;
         margin-bottom: 32px;
     }
+
     .footer__copyright{
         color: #A0A0A0;
     }
