+++ src/components/Header/Header.js
@@ -3,7 +3,6 @@ import './Header.css';
 import React from 'react';
 import Navigation from '../Navigation/Navigation';
 import { Link } from 'react-router-dom';
-// import { Link } from 'react-router-dom';
 
 export default function Header({isLogged = false}) {
     return (
@@ -12,7 +11,6 @@ export default function Header({isLogged = false}) {
                 <img src={logoPath} alt="Movies Explorer"/>
             </Link>
             <Navigation isLogged={isLogged} />
-        </header>
-            
-    )
+        </header>  
+    );
 }
