+++ src/components/Main/MainTitle/MainTitle.css
