+++ src/components/Main/Portfolio/Portfolio.css
@@ -8,7 +8,7 @@
     font-size: 18px;
     font-weight: 400;
     line-height: 20px;
-    color: rgba(160, 160, 160, 1);
+    color: #A0A0A0;
 }
 
 .portfolio__items {
@@ -26,7 +26,7 @@
     font-size: 30px;
     font-weight: 400;
     line-height: 50px;
-    border-bottom: 1px solid rgba(230, 230, 230, 1);
+    border-bottom: 1px solid #E8E8E8;
     padding-bottom: 19px;
     letter-spacing: -0.04em;
 }
