+++ src/components/Main/Promo/Promo.css
