+++ src/components/Main/Promo/Promo.js
