+++ src/components/Movies/Movies.js
@@ -1,4 +1,3 @@
-import './Movies.css';
 import MoviesCardList from '../MoviesCardList/MoviesCardList';
 import { InitialMovies } from '../../utils/movies';
