+++ src/components/MoviesCard/MoviesCard.css
@@ -96,7 +96,6 @@
         gap: 15px;
     }
     .card__block{
-        /* height: 90%; */
         display: flex;
         flex-direction: row;
         justify-content: space-between;
