+++ src/components/Profile/Profile.css
@@ -29,9 +29,9 @@
 .profile__buttons {
   padding: 0;
   margin: 0;
+  border: 0;
   max-width: 410px;
   gap: 16px;
-  border: 0;
   display: flex;
   flex-direction: column;
   align-items: center;
